XCode Cleanup
This is my best attempt at getting T2D compiling again on the latest version of XCode (13.2.1). I was successful in getting it to compile but I was getting a runtime error. Might need more work. I also updated zLib to the latest version released in 2017.

+++ engine/source/console/codeBlock.cc
@@ -463,7 +463,7 @@ bool CodeBlock::read(StringTableEntry fileName, Stream &st)
 #ifdef TORQUE_64
          *(U64*)(code+ip) = (U64)ste;
 #else
-         code[ip] = (U32)ste;
+         code[ip] = (U32)(size_t)ste;
 #endif
       }
    }
